requirement.txt
google-generativeai
pandas-gbq
flask